Refactor model loading and training configurations; introduce Adapter class for improved model efficiency

+++ aligntune/utils/load_model.py
@@ -27,7 +27,7 @@ def load_hf_model(
         config = PaliGemmaConfig(**model_config_file)
 
     # Create the model using the configuration
-    model = PaliGemmaForConditionalGeneration(config).to(device, dtype=torch.float16)
+    model = PaliGemmaForConditionalGeneration(config).to(device)
 
     # Load the state dict of the model
     model.load_state_dict(tensors, strict=False)
